Autorización de endpoints basada en roles de la forma no recomendada usando metadata.
