+++ src/auth/auth.controller.ts
@@ -5,7 +5,7 @@ import { AuthService } from './auth.service';
 import { RawHeaders, GetUser } from './decorators';
 import { CreateUserDto, LoginUserDto } from './dto';
 import { User } from './entities/user.entity';
-import { UserRoleGuard } from './guards/user-role/user-role.guard';
+import { UserRoleGuard } from './guards/user-role.guard';
 
 @Controller('auth')
 export class AuthController {
