--- src/auth/guards/user-role/user-role.guard.ts
@@ -1,12 +1,0 @@
-import { CanActivate, ExecutionContext, Injectable } from '@nestjs/common';
-import { Observable } from 'rxjs';
-
-@Injectable()
-export class UserRoleGuard implements CanActivate {
-  canActivate(
-    context: ExecutionContext,
-  ): boolean | Promise<boolean> | Observable<boolean> {
-    console.log('UserRoleGuard');
-    return true;
-  }
-}
